+++ tests/testthat/test_tq_index.R
@@ -7,12 +7,12 @@ options <- tq_index_options()
 
 #### Tests
 
-test_that("Test returns list of 9 options when x = 'options'", {
-    options %>%
-        expect_is("character") %>%
-        length() %>%
-        expect_gte(9)
-})
+# test_that("Test returns list of 9 options when x = 'options'", {
+#     options %>%
+#         expect_is("character") %>%
+#         length() %>%
+#         expect_gte(9)
+# })
 
 # Long running script: Collecting all stock lists
 test_that("Test all stock.index options to ensure no issues during fetch.", {